Move macros to a module in the double crate
